+++ src/mem/ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: lvan-gef <lvan-gef@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/06 15:33:07 by lvan-gef      #+#    #+#                 */
-/*   Updated: 2024/07/07 16:08:06 by lvan-gef      ########   odam.nl         */
+/*   Updated: 2024/11/25 21:59:17 by lvan-gef      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,5 +15,5 @@
 void	ft_bzero(void *s, size_t c)
 {
 	if (c > 0)
-		ft_memset(s, 0, c);
+		ft_memset(s, 0, (int)c);
 }
