+++ src/str/ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: lvan-gef <lvan-gef@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/31 22:20:28 by lvan-gef      #+#    #+#                 */
-/*   Updated: 2024/07/07 16:40:39 by lvan-gef      ########   odam.nl         */
+/*   Updated: 2024/11/25 22:09:14 by lvan-gef      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
 	unsigned int	index;
-	int				str_len;
+	size_t			str_len;
 	char			*str;
 
 	if (!s)
